EventBrite_Clone_Backend-SW-Project-Backend/pyproject.toml
[tool.poetry]
name = "fine"
version = "0.1.0"
description = ""
authors = ["Ismail Tawfik <elshelyziad2gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7.9"
djangorestframework = "^3.14.0"
djongo = "^1.3.6"
pymongo = "3.12.3"
django = "3.0.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
